+++ backend/app/scripts/import_daily_price_bundle.py
@@ -8,7 +8,11 @@ from pathlib import Path
 
 from app.database import SessionLocal
 from app.scripts._runtime import prepare_runtime
-from app.services.daily_price_bundle_service import DailyPriceBundleService
+from app.services.daily_price_bundle_contract import (
+    DAILY_PRICE_BAR_PERIOD,
+    DAILY_PRICE_MANIFEST_SCHEMA_VERSION,
+    expected_bundle_metadata_from_manifest,
+)
 from app.utils.file_hashing import sha256_file
 from app.wiring.bootstrap import get_daily_price_bundle_service
 
@@ -22,8 +26,7 @@ def verify_daily_price_bundle_manifest(
     expected_bundle_asset_name: str | None = None,
 ) -> dict:
     manifest = json.loads(manifest_path.read_text(encoding="utf-8"))
-    expected_schema = DailyPriceBundleService.DAILY_PRICE_MANIFEST_SCHEMA_VERSION
-    if manifest.get("schema_version") != expected_schema:
+    if manifest.get("schema_version") != DAILY_PRICE_MANIFEST_SCHEMA_VERSION:
         raise ValueError(
             "Unsupported daily price manifest schema version: "
             f"{manifest.get('schema_version')!r}"
@@ -32,7 +35,7 @@ def verify_daily_price_bundle_manifest(
     bundle_asset_name = expected_bundle_asset_name or input_path.name
     expected_values = {
         "bundle_asset_name": bundle_asset_name,
-        "bar_period": DailyPriceBundleService.DAILY_PRICE_BAR_PERIOD,
+        "bar_period": DAILY_PRICE_BAR_PERIOD,
     }
     if expected_market is not None:
         expected_values["market"] = expected_market.upper()
@@ -106,9 +109,7 @@ def main() -> int:
             expected_as_of_date=args.expected_as_of_date,
             expected_bundle_asset_name=args.expected_bundle_asset_name,
         )
-        expected_metadata = DailyPriceBundleService.expected_bundle_metadata_from_manifest(
-            manifest
-        )
+        expected_metadata = expected_bundle_metadata_from_manifest(manifest)
 
     with SessionLocal() as db:
         stats = service.import_daily_price_bundle(
