+++ backend/tests/unit/test_daily_price_bundle_service.py
@@ -2,72 +2,20 @@ from __future__ import annotations
 
 import json
 from datetime import date
-from types import SimpleNamespace
 from unittest.mock import MagicMock
 
 import pytest
-from sqlalchemy import create_engine
-from sqlalchemy.orm import sessionmaker
 
-from app.database import Base
 from app.models.app_settings import AppSetting
 from app.models.stock import StockPrice
-from app.models.stock_universe import StockUniverse, UNIVERSE_STATUS_ACTIVE
-from app.services.daily_price_bundle_reader import (
-    StreamingJsonReader,
-    read_daily_price_bundle_metadata,
-)
 from app.services.daily_price_bundle_service import DailyPriceBundleService
 
-
-def _make_session():
-    engine = create_engine("sqlite:///:memory:")
-    Base.metadata.create_all(engine)
-    return sessionmaker(bind=engine, autocommit=False, autoflush=False)
-
-
-def _stock_row(symbol: str, market: str, exchange: str, market_cap: float) -> StockUniverse:
-    return StockUniverse(
-        symbol=symbol,
-        market=market,
-        exchange=exchange,
-        is_active=True,
-        status=UNIVERSE_STATUS_ACTIVE,
-        status_reason="active",
-        market_cap=market_cap,
-    )
-
-
-def _price_row(symbol: str, day: date, close: float) -> StockPrice:
-    return StockPrice(
-        symbol=symbol,
-        date=day,
-        open=close - 1,
-        high=close + 1,
-        low=close - 2,
-        close=close,
-        adj_close=close - 0.5,
-        volume=1_000_000,
-    )
-
-
-def _bundle_price(
-    *,
-    day: str = "2026-04-18",
-    open_price: float = 100.0,
-    high: float = 101.0,
-    low: float = 99.0,
-    close: float = 100.5,
-) -> dict[str, object]:
-    return {
-        "date": day,
-        "open": open_price,
-        "high": high,
-        "low": low,
-        "close": close,
-        "adj_close": close,
-        "volume": 1_000_000,
-    }
+from .daily_price_bundle_test_helpers import (
+    bundle_price as _bundle_price,
+    make_session as _make_session,
+    price_row as _price_row,
+    stock_row as _stock_row,
+)
 
 
 def _make_service(session_factory):
@@ -363,124 +311,6 @@ def test_export_daily_price_bundle_rejects_non_finite_coverage_threshold(tmp_pat
     db.close()
 
 
-def test_sync_from_github_up_to_date_exposes_manifest_metadata():
-    session_factory = _make_session()
-    db = session_factory()
-    service = _make_service(session_factory)
-
-    result = service.sync_from_github(
-        db,
-        market="US",
-        github_sync_service=SimpleNamespace(
-            fetch_latest_bundle=lambda **kwargs: {
-                "status": "up_to_date",
-                "manifest": {
-                    "market": "US",
-                    "as_of_date": "2026-04-18",
-                    "source_revision": "daily_prices_us:20260418120000",
-                    "bundle_asset_name": "daily-price-us-20260418.json.gz",
-                    "bar_period": "2y",
-                    "symbol_count": 2,
-                },
-                "bundle_path": None,
-                "bundle_asset_name": "daily-price-us-20260418.json.gz",
-                "source_revision": "daily_prices_us:20260418120000",
-            }
-        ),
-    )
-
-    assert result["status"] == "up_to_date"
-    assert result["as_of_date"] == "2026-04-18"
-    assert result["bar_period"] == "2y"
-    assert result["symbol_count"] == 2
-    db.close()
-
-
-def test_sync_from_github_passes_allow_stale_to_release_sync_service():
-    session_factory = _make_session()
-    db = session_factory()
-    service = _make_service(session_factory)
-    captured_kwargs = {}
-
-    def _fetch_latest_bundle(**kwargs):
-        captured_kwargs.update(kwargs)
-        return {
-            "status": "missing_manifest",
-            "manifest": None,
-            "bundle_path": None,
-            "bundle_asset_name": None,
-            "source_revision": None,
-        }
-
-    result = service.sync_from_github(
-        db,
-        market="US",
-        allow_stale=True,
-        github_sync_service=SimpleNamespace(fetch_latest_bundle=_fetch_latest_bundle),
-    )
-
-    assert result["status"] == "missing_manifest"
-    assert captured_kwargs["allow_stale"] is True
-    db.close()
-
-
-def test_sync_from_github_rejects_manifest_market_mismatch():
-    session_factory = _make_session()
-    db = session_factory()
-    service = _make_service(session_factory)
-
-    result = service.sync_from_github(
-        db,
-        market="US",
-        github_sync_service=SimpleNamespace(
-            fetch_latest_bundle=lambda **kwargs: {
-                "status": "up_to_date",
-                "manifest": {
-                    "market": "HK",
-                    "as_of_date": "2026-04-18",
-                    "source_revision": "daily_prices_hk:20260418120000",
-                    "bundle_asset_name": "daily-price-hk-20260418.json.gz",
-                    "bar_period": "2y",
-                    "symbol_count": 2,
-                },
-                "bundle_path": None,
-                "bundle_asset_name": "daily-price-hk-20260418.json.gz",
-                "source_revision": "daily_prices_hk:20260418120000",
-            }
-        ),
-    )
-
-    assert result["status"] == "invalid_manifest"
-    assert "does not match requested market" in str(result["error"])
-    db.close()
-
-
-def test_daily_price_bundle_metadata_reads_scalar_across_small_chunks(
-    monkeypatch,
-    tmp_path,
-):
-    monkeypatch.setattr(StreamingJsonReader, "_CHUNK_SIZE", 3)
-    bundle_path = tmp_path / "daily-price-us.json"
-    bundle_path.write_text(
-        json.dumps(
-            {
-                "schema_version": DailyPriceBundleService.DAILY_PRICE_BUNDLE_SCHEMA_VERSION,
-                "market": "US",
-                "as_of_date": "2026-04-18",
-                "bar_period": DailyPriceBundleService.DAILY_PRICE_BAR_PERIOD,
-                "source_revision": "daily_prices_us:20260418120000",
-                "symbol_count": 123456,
-                "rows": [],
-            }
-        ),
-        encoding="utf-8",
-    )
-
-    metadata = read_daily_price_bundle_metadata(bundle_path)
-
-    assert metadata["symbol_count"] == 123456
-
-
 def test_import_daily_price_bundle_skips_redis_warm_for_two_year_bundle(tmp_path):
     session_factory = _make_session()
     db = session_factory()
@@ -735,7 +565,6 @@ def test_import_daily_price_bundle_rejects_invalid_ohlcv_and_rolls_back(tmp_path
     assert db.query(StockPrice).filter(StockPrice.symbol == "AAPL").count() == 0
     db.close()
 
-
 def test_import_daily_price_bundle_rejects_duplicate_symbols_and_rolls_back(
     monkeypatch,
     tmp_path,
@@ -866,146 +695,4 @@ def test_import_daily_price_bundle_rolls_back_state_on_persist_error(
 
     assert db.query(AppSetting).count() == 0
     assert db.query(StockPrice).filter(StockPrice.symbol == "AAPL").count() == 0
-    db.close()
-
-
-def test_sync_from_github_import_uses_manifest_metadata_without_payload_prescan(
-    monkeypatch,
-    tmp_path,
-):
-    session_factory = _make_session()
-    db = session_factory()
-    db.add(_stock_row("AAPL", "US", "NASDAQ", 1000.0))
-    db.commit()
-
-    service = DailyPriceBundleService()
-    monkeypatch.setattr(
-        service,
-        "_read_bundle_metadata",
-        MagicMock(side_effect=AssertionError("payload metadata prescan")),
-    )
-    bundle_path = tmp_path / "daily-price-us.json"
-    bundle_path.write_text(
-        json.dumps(
-            {
-                "rows": [
-                    {
-                        "symbol": "AAPL",
-                        "prices": [
-                            {
-                                "date": "2026-04-18",
-                                "open": 100.0,
-                                "high": 101.0,
-                                "low": 99.0,
-                                "close": 100.5,
-                                "adj_close": 100.0,
-                                "volume": 1_000_000,
-                            }
-                        ],
-                    }
-                ],
-                "schema_version": service.DAILY_PRICE_BUNDLE_SCHEMA_VERSION,
-                "market": "US",
-                "as_of_date": "2026-04-18",
-                "bar_period": service.DAILY_PRICE_BAR_PERIOD,
-                "source_revision": "daily_prices_us:20260418120000",
-                "symbol_count": 1,
-            }
-        ),
-        encoding="utf-8",
-    )
-    manifest = {
-        "market": "US",
-        "as_of_date": "2026-04-18",
-        "source_revision": "daily_prices_us:20260418120000",
-        "bundle_asset_name": bundle_path.name,
-        "bar_period": service.DAILY_PRICE_BAR_PERIOD,
-        "symbol_count": 1,
-    }
-
-    result = service.sync_from_github(
-        db,
-        market="US",
-        github_sync_service=SimpleNamespace(
-            fetch_latest_bundle=lambda **kwargs: {
-                "status": "success",
-                "manifest": manifest,
-                "bundle_path": str(bundle_path),
-                "bundle_asset_name": bundle_path.name,
-                "source_revision": manifest["source_revision"],
-            }
-        ),
-    )
-
-    assert result["status"] == "success"
-    assert result["imported_symbols"] == 1
-    assert db.query(StockPrice).filter(StockPrice.symbol == "AAPL").count() == 1
-    service._read_bundle_metadata.assert_not_called()
-    db.close()
-
-
-def test_sync_from_github_rejects_bundle_metadata_mismatch_and_rolls_back(
-    tmp_path,
-):
-    session_factory = _make_session()
-    db = session_factory()
-
-    service = DailyPriceBundleService()
-    bundle_path = tmp_path / "daily-price-us.json"
-    bundle_path.write_text(
-        json.dumps(
-            {
-                "rows": [
-                    {
-                        "symbol": "AAPL",
-                        "prices": [
-                            {
-                                "date": "2026-04-18",
-                                "open": 100.0,
-                                "high": 101.0,
-                                "low": 99.0,
-                                "close": 100.5,
-                                "adj_close": 100.0,
-                                "volume": 1_000_000,
-                            }
-                        ],
-                    }
-                ],
-                "schema_version": service.DAILY_PRICE_BUNDLE_SCHEMA_VERSION,
-                "market": "HK",
-                "as_of_date": "2026-04-18",
-                "bar_period": service.DAILY_PRICE_BAR_PERIOD,
-                "source_revision": "daily_prices_hk:20260418120000",
-                "symbol_count": 1,
-            }
-        ),
-        encoding="utf-8",
-    )
-    manifest = {
-        "market": "US",
-        "as_of_date": "2026-04-18",
-        "source_revision": "daily_prices_us:20260418120000",
-        "bundle_asset_name": bundle_path.name,
-        "bar_period": service.DAILY_PRICE_BAR_PERIOD,
-        "symbol_count": 1,
-    }
-
-    with pytest.raises(ValueError, match="bundle market 'HK' does not match manifest 'US'"):
-        service.sync_from_github(
-            db,
-            market="US",
-            github_sync_service=SimpleNamespace(
-                fetch_latest_bundle=lambda **kwargs: {
-                    "status": "success",
-                    "manifest": manifest,
-                    "bundle_path": str(bundle_path),
-                    "bundle_asset_name": bundle_path.name,
-                    "source_revision": manifest["source_revision"],
-                }
-            ),
-        )
-
-    assert db.query(StockPrice).count() == 0
-    assert db.query(AppSetting).count() == 0
-    assert not bundle_path.exists()
     db.close()
